Displaying ChatGPT error
When we use a non-subscribe token. which is chatgpt-3.5 model token, then it just shows a blurred loader. Or if I mistakenly type the wrong token, it still shows a blurred loader.

In this commit, I am showing a message, so that the user would look back at its API Token.